move sdk merge into a separate file
